MADLIBS GENERATOR
This is a simple code in python for  Mad Libs Generator! A Mad Libs Generator is a fun, interactive game where you fill in the blanks of a story with various types of words (like nouns, verbs, adjectives, etc.) Perfect for a laugh with friends or to spark some creative storytelling.